[Fix] 여행 준비물 체크 페이지 내 아코디언 기능 추가 #22 - 체킹 관련 handler 코드를 아코디언에 배치 - state로 아코디언 기능 적용 - 기존 section에 아코디언 배치
